requirements.txt
# requirements.txt
requests
beautifulsoup4
python-telegram-bot==20.8